pyproject.toml
[project]
name = "uni-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Francesco Filicetti",email = "francesco.filicetti@unical.it"}
]
license = "AGPL-3.0-or-later"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.10,<7.0",
    "djangorestframework>=3.16.1,<4.0",
    "djangorestframework-simplejwt>=5.5.1,<6.0.0",
    "drf-spectacular>=0.29.0,<1.0",
    "requests>=2.32.5,<3.0",
    "uni-esse3api",
]

[dependency-groups]
prod = [
    "mysqlclient>=2.2.7,<3.0",
    "redis>=7.1.0,<8.0",
    "uwsgi>=2.0.31,<3.0",
]

[tool.uv.sources]
uni-esse3api = { git = "https://github.com/UniversitaDellaCalabria/uniEsse3Api.git" }
